first version of templated serialization and verification functions
